European Bank’s Customer Satisfaction Analysis & Churn Prediction
Customer churn is a critical issue for businesses in competitive markets. Understanding and predicting churn can help companies strategize proactive measures on customer retention, improve customer satisfaction, and increase revenue while acquiring new customers. The aim is to predict customer churn (whether a customer will exit the bank or not). This is a binary classification problem where the target variable is Exited.